requirements.txt
cffi==1.15.1
cryptography==37.0.4
pycparser==2.21
PyJWT==2.4.0
